Add tests to reproduce the merge problem.
The problem append when generate_dict handle a ranged key.

If the ranged_key is handle first the bug is not visible.
In case the range is handled after others keys, the other keys
values is ignored.

This bug depend of your key name example :

```>>> a={'host10': 'foo', '=host10-12': 'foo'}
>>> b={'node1': 'foo', '=node1-3': 'foo'}
>>> a
{'=host10-12': 'foo', 'host10': 'foo'}
>>> b
{'node1': 'foo', '=node1-3': 'foo'}
```

For `hostX` key the bug will not be visible. For `nodeX` the bug is visible.