detect: Adding more serial id fallbacks
As per issue #206, we have systems where serial numbers are .... weird.
We have system that are even much more evil like DMI not providing any
unique ID at all...

This patch is trying to solve all that different cases.

If the serial is known to be buggy then we search for a possible
fallback in the following order :
- system uuid
- motherboard serial
- MAC address of the first nic

Note that the system uuid will be used unless it is known to be buggy or empty.

If a system doesn't provide any valid 'product SN' or 'product uuid' or
'motherboard serial' (yes this does exist), then we fallback on the last
unique ID we have on a system which is the mac address of one nic : this
is not optimal but at least that makes the installation process staying
alive. That's really a workaround.