Add support for benchmark tests
This patch is adding support for benchmark tests as part of the
hardware-detect tool.

The test is optional and can be activated by passing the --benchmark
parameter to the tool. The parameter supports passing multiple values,
acceptable values right now are "disk", "cpu" and "mem".

Another parameter --benchmark-disk-destructive have been added and if
set it will turn on the destructive mode for the disk benchmark.

The new help menu:

usage: hardware-detect [-h] [-H] [--benchmark component [component ...]]
                       [--benchmark-disk-destructive]

optional arguments:
  -h, --help            show this help message and exit
  -H, --human           Print output in human readable format
  --benchmark component [component ...], -b component [component ...]
                        Run benchmark for specific components. Valid
                        components are: disk, cpu, mem
  --benchmark-disk-destructive
                        If specified make the disk component benchmark
			to be destructive