Remove the psutil dependency
psutil was used just to get the amount of available memory in the system,
psutil makes it nice to be able to do it in multiple platforms like bsd,
linux and windows. But since we only run this code on linux we can get
this information directly from /proc/meminfo so that we don't need to
add an extra dependency.