detect: Automatic megacli's LD description
When reading the Logical Drive properties, the previous code was pretty
long and assumed the name of the properties to manipulate.

That was a pain as some properties were present sometimes but not on
some other.

This patch just make the code easier to read, more robust and reports
much more info as we had previously.

The principle is very easy to understand .... "just report what we have"....

That should have been like that since a while.