megacli: Searching various version of megacli executable
Depending on the version or the distribution method, megacli could have
various casting like megacli or MegaCli.

This patch does add a method to find which one to use.

As we still run on python2.7, this patch reuse which() from python3 and
an homemade search_exec() function to find all possible name of the same
exec.

Therefore it's now possible to detect both of them and use it. We had a
case where megacli commands were not executed because of the name
cast...

Patch tested on an hypervisor running MegaCli.