cardiff: Comparing hardware physical disks
Just discovered that no code was calling the hardware physical disks
comparaison while it's one of the key component on a server.

This patch is simply making the proper call to get hpa & megacli PD
reporting.

It also add a couple of unique IDs that are reported by the controller
but that should not be taken in consideration when comparing the
physical disks (PDs).