cardiff: Tests shall use check.xxxx() functions
The testing code was using a copy of the check.*() functions. That was bad as
any modification in the check code was not integrated naturally in the test
code.

Since commit 8d87d433c0481913cef43a775995b0064ddecf22, it's now possible to call
directly the check.*() functions from the test code.

This patch only take care of replacing the current duplicated code by the proper
check.*() call.