detect: Moving disk's information order
This patch is about inverting some disk info to ease making proper edeploy's
role with it.

The actual code was printing disk's info in the following order :

    size
    vendor
    model
    <other>

When reusing an hardware output as a edpeloy role, that's an issue and edeploy
requires the top discriminating element first. This patch just invert the order
like :

    model
    vendor
    <other>
    size

This way turning an hw report to an edeploy role is much easier.