matcher: Adding range() function
This patch add a new spec helper called 'range'

It does allow putting a matching rule to check if a value is part of a range.

The prototype is like range(lower_bound, upper_bound) and the code return True
if (lower_bound <= x <= upper_bound).

A typical usage would be :

 ('ipmi', '+12 V', 'value', 'range(11.8, 12.2)'),