Fix Infiniband device enumeration
Previously, enumeration of Infiniband devices in
infiniband.ib_card_drv() resulted in an empty last entry in the list,
due to the use of split() with a trailing newline. This resulted in a
KeyError exception in the detect.detect_infiniband() function.

This change removes the empty last entry from the device list.