Fix Infiniband device naming
The detect.detect_infiniband() function causes all Infiniband cards to
be indexed as card0. This is because the variable ib_card is not
incremented on each iteration of the loop (devices is). This change
removes the redundant devices loop variable, instead using ib_card for
the loop index and card index.