Do not access /sys on system running unit tests
While adding tests for the ppc64le[1] lscpu issues it was clear that we're
access /sys on the host system  This means hat the unit tests will fail
depending on what host tox is run on.

Ensure that os.path.exists returns false to avoid that issue

We need to be a little careful with how we check the mocked calls.
assert_has_calls()  just ensures that all the calls in the expected
argument happen in order but doesn't fail if there are additional calls
between.  So comparing the lists are equal is much stricter

[1] Later in the series