Be more robust when parseing NUMA information
Some Architectures can have sparse and empty NUMA nodes so don't assume
that if we have 4 NUMA nodes they will be called 0,1,2 and 3.

Also ensure we have CPUs in a NUMA node before trying to detect the
mapping.

Note: Unit tests don't need to be changed so this is safe given out
existing data