Don't repeat the same string formatting
Repeating "physical_{}".format(processor) leads to longer lines that
need breaking.  Rather than repeat that string operation save it in a
temporary variable and use that.

We don't strictly need to do the same thing for NUMA and logical tags but
to keep the sections following the same code style we do it there too.