detect: Refactor the code in get_cpus()
The body of get_cpus() adds items to a list and the items may require
type enforcement.  When adding support for ppc64le[1] we need to check
if keys exist, and allow for some elements having different types
(int[x86_64] and str[ppc64le]).

To handle all that exract the common code into a loop body and iterate
over a list of tuples

[1] Later in the series