Basic support for nvme devices
Adding support for detecting nvme devices based on device names.
With this patch we can read basic information on nvme devices
including and not limited to name, size, vendor and model.

Also splitting detect_disks under detect.py in multiple functions
under diskinfo.

This has been successfully tested on a system with a single nvme
device and a system with non-nvme devices.