Remove hardware.cardiff and the hardware-cardiff tool (fixes #124)
These parts of the library are no longer maintained and bring in
heavy dependencies on numpy and pandas (pulling in X11 libraries).

It's also unclear if anybody is using it nowadays. If somebody
complains, we can always resurrect this code as a new Python package,
since there are no dependencies between it and the rest of the library.