detect: Fixing numa cpu_count computation
The actual code was confused by a syntax like :  0,1,16,17
It was considering each value as a number of cores reporting 34 to this example.

This patch modify the logic and count only one core per comma when there isn't any dash ("-") in it.
This add a sample of an AMD 7302 so have this configuration.

Signed-off-by: Erwan Velu <e.velu@criteo.com>