Fix final output
Since the new detect functions in the separate modules return
lists, we should not append them to the final list but instead
extend it.
This fixes the issue described in [1] where the extra_hardware
plugin in ironic-inspector tries to convert the data collected
by the ipa ramdisk but gets an unexpected list inside a list.

[1] https://bugs.launchpad.net/tripleo/+bug/1901917