Add line to setup.cfg limiting to Python >=3.6
This adds "Requires-Python: >=3.6" to the PKG-INFO file, which will then
in turn be processed by PyPI and pip and prevent users on Python 2 from
being able to `pip install hardware`.

Also update metadata keys to use underscore since recent versions of
setuptools are warning that usage of dash-separated keys will not be
supported in future versions.