fix: use `--random_generator=tausworthe64`
If disks are large enough then FIO will write to stdout a warning that
the default random generator is not sufficient. As this is written to
stdout the subsequent attempt to parse the JSON will fail as it contains
the warning before the JSON object.

To fix we can set the random generator to `tausworthe64` which will
suppress the warning allow for the JSON to be parsed without error.

This issue has been noticed on disks around 64TB in size however could
occur on smaller disks.